split files and move to right_border_tabs
